+++ apps/site/src/app/admin/directors/organizers/page.tsx
@@ -1,0 +1,1 @@
+export { default as default } from "./Organizers";
